(many):	remove const spec from FindVarMethod. That was wrong
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@2586 06827809-a52a-0410-b366-d66718629ded

+++ agent/snmpd.c
@@ -750,7 +750,7 @@ main(int argc, char *argv[])
         perror("fopen");
         exit(1);
       }
-      fprintf(PID, "%d\n", getpid());
+      fprintf(PID, "%d\n", (int)getpid());
       fclose(PID);
     }
